Move building into pyast directory
